Split off config_eeprom.c
