+++ Makefile
@@ -372,6 +372,7 @@ COMMON_SRC = \
             common/streambuf.c \
             config/config.c \
             config/feature.c \
+            config/config_eeprom.c \
             fc/runtime_config.c \
             drivers/logging.c \
             drivers/adc.c \
